change long long to i64 change segment struct name

+++ binomial coefficient.hpp
@@ -3,8 +3,8 @@ std::vector<i32> pw(N+1) , pwinv(N+1);
 i32 fp(i32 a, i32 n, i32 mod) {
     i32 res = 1;
     while (n) {
-        if (n % 2) res = (long long)res * a % mod;
-        a = (long long)a * a % mod;
+        if (n % 2) res = (i64)res * a % mod;
+        a = (i64)a * a % mod;
         n >>= 1;
     }
     return res;
@@ -14,10 +14,10 @@ i32 C(i32 a, i32 b) {
     if (pw[0] == 0) {
         pwinv[0] = pw[0] = 1;
         for (i32 i = 1;i <= N;i++) {
-            pw[i] = (long long)pw[i - 1] * i % mod;
+            pw[i] = (i64)pw[i - 1] * i % mod;
             pwinv[i] = inv(pw[i]);
         }
     }
     if (a < b) return 0;
-    return (long long)pw[a] * pwinv[b] % mod * pwinv[a-b] % mod;
+    return (i64)pw[a] * pwinv[b] % mod * pwinv[a-b] % mod;
 }
